fix: refactor repeated function
