+++ src/views/ChatView.vue
@@ -569,6 +569,7 @@
 import Stomp from 'webstomp-client'
 import SockJS from 'sockjs-client'
 import { mapGetters } from 'vuex';
+import { getAddr } from '../utils';
     
 export default {
   data() {
@@ -747,34 +748,7 @@ export default {
     handleLogout() {
       this.$store.dispatch('logout');
     },
-    addKakaoMapScript() {
-      const script = document.createElement("script");
-      /* global kakao */
-      script.onload = () => kakao.maps.load(this.initMap);
-      script.src =
-        "https://dapi.kakao.com/v2/maps/sdk.js?autoload=false&appkey=4e77d9b3460eb3b942634fb28e5e1c40&libraries=services";
-      document.head.appendChild(script);
-    },
-    getAddr() {
-      if (navigator.geolocation) {
-        navigator.geolocation.getCurrentPosition(position => {
-          const lat = position.coords.latitude;
-          const lng = position.coords.longitude;
-  
-          let geocoder = new kakao.maps.services.Geocoder();
-          let coord = new kakao.maps.LatLng(lat, lng);
-          let callback = (result, status) => {
-            if (status === kakao.maps.services.Status.OK) {
-              console.log(result[0].road_address.region_1depth_name);
-              this.$store.state.currentLocation = result[0].road_address.region_1depth_name
-            }
-          };
-          geocoder.coord2Address(coord.getLng(), coord.getLat(), callback);
-        });
-      } else {
-        console.log("Geolocation is not supported by this browser.");
-      }
-    },
+    getAddr,
     loginToken() {
       const token = this.getToken; // Vuex 스토어에서 토큰 값을 가져옴
