+++ src/views/PostWriteView.vue
@@ -2,466 +2,6 @@
 import AppFooter from '@/common/AppFooter.vue';
 </script>
 <template>
-  <!--시스템 바-->
-  <v-app-bar
-    height="80"
-    :elevation="1"
-  >
-    <v-container>
-      <v-row align="center">
-        <v-col cols="2">
-          <router-link to="/">
-            <v-img
-              src="../img/logo.png"
-              :style="{ 'max-height': '100px' }"
-            />
-          </router-link>
-        </v-col>
-        <v-col cols="5">
-          <v-tabs
-            v-model="active_tab"
-            color="#5E913B"
-          >
-            <router-link
-              to="/"
-              style="text-decoration: none; color:black;"
-            >
-              <v-tab @click="$store.state.myGlobalVariable = 0">
-                <p class="font-weight-bold text-h6 mx-4">
-                  {{ link[0] }}
-                </p>
-              </v-tab>
-            </router-link>
-            <router-link
-              to="/board/ALL/FREE"
-              style="text-decoration: none; color:black;"
-              @click="$store.state.boardCheck = 0"
-            >
-              <v-menu open-on-hover>
-                <template #activator="{ props }">
-                  <v-tab v-bind="props">
-                    <p class="font-weight-bold text-h6 mx-4">
-                      {{ link[1] }}
-                    </p>
-                  </v-tab>
-                </template>
-                <v-list>
-                  <v-list-item align="center">
-                    <router-link
-                      to="/board/ALL/FREE"
-                      style="text-decoration: none; color:black;"
-                      @click="region_all"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.boardCheck = 0"
-                      >
-                        자유
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/board/SEOUL/TALK"
-                      style="text-decoration: none; color:black;"
-                      @click="region_seoul"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.boardCheck = 1"
-                      >
-                        서울
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/board/PUSAN/TALK"
-                      style="text-decoration: none; color:black;"
-                      @click="region_busan"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.boardCheck = 2"
-                      >
-                        부산
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/board/ULSAN/TALK"
-                      style="text-decoration: none; color:black;"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.boardCheck = 3"
-                      >
-                        울산
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/board/KEYNONGNAM/TALK"
-                      style="text-decoration: none; color:black;"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.boardCheck = 4"
-                      >
-                        경남
-                      </v-list-item-title>
-                    </router-link>
-                  </v-list-item>
-                </v-list>
-              </v-menu>
-            </router-link>
-            <router-link
-              to="/independent/CLEAN"
-              style="text-decoration: none; color:black;"
-              @click="$store.state.independentCheck = 0"
-            >
-              <v-menu open-on-hover>
-                <template #activator="{ props }">
-                  <v-tab v-bind="props">
-                    <p class="font-weight-bold text-h6 mx-4">
-                      {{ link[2] }}
-                    </p>
-                  </v-tab>
-                </template>
-                <v-list>
-                  <v-list-item align="center">
-                    <router-link
-                      to="/independent/CLEAN"
-                      style="text-decoration: none; color:black;"
-                      @click="independent_clean"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.independentCheck = 0"
-                      >
-                        청소
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/independent/WASH"
-                      style="text-decoration: none; color:black;"
-                      @click="independent_wash"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.independentCheck = 1"
-                      >
-                        세탁
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/independent/COOK"
-                      style="text-decoration: none; color:black;"
-                      @click="independent_cook"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.independentCheck = 2"
-                      >
-                        요리
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/independent/HEALTH"
-                      style="text-decoration: none; color:black;"
-                      @click="independent_health"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.independentCheck = 3"
-                      >
-                        건강
-                      </v-list-item-title>
-                    </router-link>
-                    <v-divider
-                      :thickness="1"
-                      class="border-opacity-25 mb-2"
-                    />
-                    <router-link
-                      to="/independent/ETC"
-                      style="text-decoration: none; color:black;"
-                      @click="independent_etc"
-                    >
-                      <v-list-item-title
-                        class="my-2"
-                        @click="$store.state.independentCheck = 4"
-                      >
-                        기타
-                      </v-list-item-title>
-                    </router-link>
-                  </v-list-item>
-                </v-list>
-              </v-menu>
-            </router-link>
-          </v-tabs>
-        </v-col>
-        <v-col cols="3">
-          <v-card-text>
-            <v-text-field
-              v-model="searchText"
-              :loading="error"
-              density="compact"
-              variant="outlined"
-              label="통합검색"
-              append-inner-icon="mdi-magnify"
-              single-line
-              hide-details
-              @click:append-inner="totalSearch"
-              @keydown.enter="totalSearch"
-            />
-          </v-card-text>
-        </v-col>
-        <v-col
-          v-if="!getToken"
-          cols="1"
-        >
-          <router-link to="/login">
-            <v-btn
-              variant="flat"
-              color="#5E913B"
-              class="font-weight-bold"
-              style="width:100%; height:40px"
-            >
-              <div class="text-white">
-                로그인
-              </div>
-            </v-btn>
-          </router-link>
-        </v-col>
-        <v-col
-          v-if="!getToken"
-          cols="1"
-        >
-          <router-link to="/signup">
-            <v-btn
-              variant="flat"
-              color="#5E913B"
-              class="font-weight-bold"
-              style="width:100%; height:40px"
-            >
-              <div class="text-white">
-                회원가입
-              </div>
-            </v-btn>
-          </router-link>
-        </v-col>
-        <v-col
-          v-if="getToken"
-          cols="2"
-        >
-          <v-menu :close-on-content-click="false">
-            <template #activator="{ props }">
-              <v-row
-                align="center"
-                justify="end"
-              >
-                <v-btn
-                  v-bind="props"
-                  class="font-weight-bold mr-3"
-                  variant="tonal"
-                  color="green-lighten-1"
-                >
-                  <v-img
-                    :height="25"
-                    :width="25"
-                    src="../img/user.png"
-                    style="color:#2E471D"
-                  />
-                  <span
-                    style="color: #5E913B;"
-                    class="font-weight-bold"
-                  >{{ userNickName }}</span>
-                </v-btn>
-              </v-row>
-            </template>
-            <v-card
-              :height="showLocationAuthentication ? 400 : 275"
-              :width="250"
-            >
-              <v-list>
-                <v-list-item style="text-align: center;">
-                  <v-row
-                    class="mt-1"
-                    style="cursor: pointer;"
-                  >
-                    <v-col cols="1" />
-                    <v-col cols="auto">
-                      <v-img
-                        :height="25"
-                        :width="25"
-                        src="../img/infomation.png"
-                        class=""
-                      />
-                    </v-col>
-                    <v-col cols="2" />
-                    <v-col cols="auto">
-                      <v-list-item-title
-                        style="font-size:18px"
-                        class="font-weight-bold"
-                      >
-                        내 정보
-                      </v-list-item-title>
-                    </v-col>
-                  </v-row>
-                  <v-divider
-                    :thickness="1"
-                    class="border-opacity-25 my-5"
-                  />
-                  <v-row style="cursor: pointer;">
-                    <v-col cols="1" />
-                    <v-col cols="auto">
-                      <v-img
-                        :height="25"
-                        :width="25"
-                        src="../img/chatting.png"
-                        class=""
-                      />
-                    </v-col>
-                    <v-col cols="2" />
-                    <v-col cols="auto">
-                      <router-link
-                        :to="`/chatRooms`"
-                        style="text-decoration: none; color:black;"
-                      >
-                        <v-list-item-title
-                          style="font-size:18px"
-                          class="font-weight-bold"
-                        >
-                          채팅
-                        </v-list-item-title>
-                      </router-link>
-                    </v-col>
-                  </v-row>
-                  <v-divider
-                    :thickness="1"
-                    class="border-opacity-25 my-5"
-                  />
-                  <v-row
-                    style="cursor: pointer;"
-                    class="mb-3"
-                    @click="showLocationAuthentication = !showLocationAuthentication"
-                  >
-                    <v-col cols="1" />
-                    <v-col cols="auto">
-                      <v-img
-                        :height="25"
-                        :width="25"
-                        src="../img/location.png"
-                        class=""
-                      />
-                    </v-col>
-                    <v-col cols="2" />
-                    <v-col cols="auto">
-                      <v-list-item-title
-                        style="font-size:18px"
-                        class="font-weight-bold"
-                      >
-                        위치인증
-                      </v-list-item-title>
-                    </v-col>
-                  </v-row>
-
-                  <v-row v-if="showLocationAuthentication">
-                    <v-col cols="12">
-                      <v-sheet>
-                        <div
-                          style="text-align:center; color: gray; font-size:14px"
-                          class="font-weight-bold mb-1"
-                        >
-                          현재위치불러오기
-                        </div>
-                        <v-row>
-                          <v-col cols="auto">
-                            <v-switch
-                              v-model="boolAuthentication"
-                              class="ml-3"
-                              color="success"
-                              @change="toggleLocationAuthentication"
-                            />
-                          </v-col>
-                          <v-col
-                            cols="auto"
-                            align="end"
-                            justify="end"
-                          >
-                            <div v-if="$store.state.locationAuthentication">
-                              <div class="mt-4 font-weight-bold">
-                                현재위치 : <span style="color: #5E913B">{{
-                                  $store.state.currentLocation }}</span>
-                              </div>
-                            </div>
-                            <div v-else>
-                              <div class="mt-4 font-weight-bold">
-                                현재위치 : <span style="color: #5E913B">인증필요</span>
-                              </div>
-                            </div>
-                          </v-col>
-                        </v-row>
-                      </v-sheet>
-                    </v-col>
-                  </v-row>
-
-                  <v-divider
-                    :thickness="1"
-                    class="border-opacity-25 mb-5"
-                  />
-                  <v-row style="cursor: pointer;">
-                    <v-col cols="1" />
-                    <v-col cols="auto">
-                      <v-img
-                        :height="25"
-                        :width="25"
-                        src="../img/logout.png"
-                        class=""
-                      />
-                    </v-col>
-                    <v-col cols="2" />
-                    <v-col cols="auto">
-                      <v-list-item-title
-                        style="font-size:18px"
-                        class="font-weight-bold"
-                        @click="handleLogout"
-                      >
-                        로그아웃
-                      </v-list-item-title>
-                    </v-col>
-                  </v-row>
-                </v-list-item>
-              </v-list>
-            </v-card>
-          </v-menu>
-        </v-col>
-      </v-row>
-    </v-container>
-  </v-app-bar>
-
   <div>
     <div v-if="$store.state.independentCheck === 5 && $store.state.boardCheck === 0">
       {{ setBoardCheck(0) }} {{ setTypeCheck(0) }} {{ setBoardNameCheck(0) }}
@@ -905,26 +445,6 @@ export default {
     },
     cancle() {
       this.$router.go(-1)
-    },
-    toggleLocationAuthentication() {
-      this.$store.commit('toggleLocationAuthentication');
-
-      if (this.$store.state.locationAuthentication === true) {
-        this.$axios.post("/api/members/region", { region: this.$store.state.currentLocation }, {
-          headers: {
-            Authorization: this.$store.state.token, // 헤더에 토큰 추가
-          },
-        });
-      }        
-    },
-    totalSearch() {
-      if (this.searchText !== '') {
-        const query = this.searchText ? `?searchText=${encodeURIComponent(this.searchText)}` : '';
-        window.location.href = '/search' + query;
-      }
-    },
-    handleLogout() {
-      this.$store.dispatch('logout');
     },
     addKakaoMapScript() {
       const script = document.createElement("script");
